move empty input test to `tests/`
